Bug fixed in keyvalue memory.h code and improved its tests.
In group class, when the meta_type vector reaches its maximum
capacity and needs to be reallocated, the contents of the vector
are not copied correctly.
Temporary solution is to avoid vector reallocation by specifying
the maximum size when group objects are created and protecting
the push_back function with an assert.